+++ jib-gradle-plugin/src/integration-test/java/com/google/cloud/tools/jib/gradle/SingleProjectIntegrationTest.java
@@ -61,8 +61,17 @@ public class SingleProjectIntegrationTest {
     Assert.assertTrue(parsed.isAfter(before) || parsed.equals(before));
   }
 
+  private static void assertWorkingDirectory(String expected, String imageReference)
+      throws IOException, InterruptedException {
+    Assert.assertEquals(
+        expected,
+        new Command("docker", "inspect", "-f", "{{.Config.WorkingDir}}", imageReference)
+            .run()
+            .trim());
+  }
+
   /**
-   * Asserts that the test project has the required exposed ports and labels.
+   * Asserts that the test project has the required exposed ports, labels and volumes.
    *
    * @param imageReference the image to test
    * @throws IOException if the {@code docker inspect} command fails to run
@@ -71,6 +80,13 @@ public class SingleProjectIntegrationTest {
   private static void assertDockerInspect(String imageReference)
       throws IOException, InterruptedException {
     String dockerInspect = new Command("docker", "inspect", imageReference).run();
+    Assert.assertThat(
+        dockerInspect,
+        CoreMatchers.containsString(
+            "            \"Volumes\": {\n"
+                + "                \"/var/log\": {},\n"
+                + "                \"/var/log2\": {}\n"
+                + "            },"));
     Assert.assertThat(
         dockerInspect,
         CoreMatchers.containsString(
@@ -144,6 +160,7 @@ public class SingleProjectIntegrationTest {
         JibRunHelper.buildAndRun(simpleTestProject, targetImage));
     assertDockerInspect(targetImage);
     assertSimpleCreationTimeIsAfter(beforeBuild, targetImage);
+    assertWorkingDirectory("/home", targetImage);
   }
 
   @Test
@@ -154,6 +171,7 @@ public class SingleProjectIntegrationTest {
         "Hello, world. An argument.\nrwxr-xr-x\nrwxrwxrwx\nfoo\ncat\n-Xms512m\n-Xdebug\nenvvalue1\nenvvalue2\n",
         buildAndRunComplex(targetImage, "testuser2", "testpassword2", localRegistry2));
     assertSimpleCreationTimeIsAfter(beforeBuild, targetImage);
+    assertWorkingDirectory("", targetImage);
   }
 
   @Test
@@ -164,6 +182,7 @@ public class SingleProjectIntegrationTest {
         "Hello, world. An argument.\nrwxr-xr-x\nrwxrwxrwx\nfoo\ncat\n-Xms512m\n-Xdebug\nenvvalue1\nenvvalue2\n",
         buildAndRunComplex(targetImage, "testuser", "testpassword", localRegistry1));
     assertSimpleCreationTimeIsAfter(beforeBuild, targetImage);
+    assertWorkingDirectory("", targetImage);
   }
 
   @Test
@@ -175,6 +194,7 @@ public class SingleProjectIntegrationTest {
         JibRunHelper.buildToDockerDaemonAndRun(simpleTestProject, targetImage));
     assertSimpleCreationTimeIsAfter(beforeBuild, targetImage);
     assertDockerInspect(targetImage);
+    assertWorkingDirectory("/home", targetImage);
   }
 
   @Test
@@ -196,5 +216,6 @@ public class SingleProjectIntegrationTest {
         new Command("docker", "run", "--rm", targetImage).run());
     assertDockerInspect(targetImage);
     assertSimpleCreationTimeIsAfter(beforeBuild, targetImage);
+    assertWorkingDirectory("/home", targetImage);
   }
 }