+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildTarMojo.java
@@ -21,11 +21,13 @@ import com.google.cloud.tools.jib.event.DefaultEventDispatcher;
 import com.google.cloud.tools.jib.event.EventDispatcher;
 import com.google.cloud.tools.jib.filesystem.AbsoluteUnixPath;
 import com.google.cloud.tools.jib.image.InvalidImageReferenceException;
-import com.google.cloud.tools.jib.plugins.common.AppRootInvalidException;
 import com.google.cloud.tools.jib.plugins.common.BuildStepsExecutionException;
 import com.google.cloud.tools.jib.plugins.common.BuildStepsRunner;
 import com.google.cloud.tools.jib.plugins.common.HelpfulSuggestions;
 import com.google.cloud.tools.jib.plugins.common.InferredAuthRetrievalException;
+import com.google.cloud.tools.jib.plugins.common.InvalidAppRootException;
+import com.google.cloud.tools.jib.plugins.common.InvalidContainerVolumeException;
+import com.google.cloud.tools.jib.plugins.common.InvalidWorkingDirectoryException;
 import com.google.cloud.tools.jib.plugins.common.MainClassInferenceException;
 import com.google.cloud.tools.jib.plugins.common.PluginConfigurationProcessor;
 import com.google.cloud.tools.jib.plugins.common.RawConfiguration;
@@ -104,10 +106,19 @@ public class BuildTarMojo extends JibPluginConfiguration {
               helpfulSuggestions);
       getLog().info("");
 
-    } catch (AppRootInvalidException ex) {
+    } catch (InvalidAppRootException ex) {
       throw new MojoExecutionException(
-          "<container><appRoot> is not an absolute Unix-style path: " + ex.getInvalidAppRoot());
+          "<container><appRoot> is not an absolute Unix-style path: " + ex.getInvalidPathValue(),
+          ex);
+    } catch (InvalidWorkingDirectoryException ex) {
+      throw new MojoExecutionException(
+          "<container><workingDirectory> is not an absolute Unix-style path: "
+              + ex.getInvalidPathValue(),
+          ex);
 
+    } catch (InvalidContainerVolumeException ex) {
+      throw new MojoExecutionException(
+          "<container><volumes> is not an absolute Unix-style path: " + ex.getInvalidVolume(), ex);
     } catch (InvalidImageReferenceException
         | IOException
         | CacheDirectoryCreationException