+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/RawConfiguration.java
@@ -61,9 +61,13 @@ public interface RawConfiguration {
 
   Map<String, String> getLabels();
 
+  List<String> getVolumes();
+
   List<String> getPorts();
 
   Optional<String> getUser();
+
+  Optional<String> getWorkingDirectory();
 
   boolean getUseCurrentTimestamp();